Lab02
Lab02 presents an engaging C++ programming exercise that focuses on the application of conditional statements, particularly the "if-else" construct. Participants explore the practical usage of conditional logic to control program flow and make decisions based on specified conditions. Through hands-on exercises and problem-solving tasks, this lab encourages learners to master the art of designing algorithms with conditional statements, fostering a deeper understanding of control structures in C++ programming.